refactor: seperate auth config with new file
